% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies_function.R
\name{cookies_function}
\alias{cookies_function}
\title{A cookie function}
\usage{
cookies_function(love)
}
\arguments{
\item{love}{- Recieve TRUE or FALSE}
}
\description{
This function allows you to express your love for cookies
}
\examples{
cat_function()
}
\keyword{cats}
